SakService har altfor mykje ansvar, og det å finne eining for person er ikkje opplagt at høyrer heime der. Trekkjer over til EnhetService, som kjens meir naturleg
