+++ apps/etterlatte-behandling/src/main/kotlin/behandling/EnhetService.kt
@@ -1,7 +1,16 @@
 package no.nav.etterlatte.behandling
 
+import no.nav.etterlatte.behandling.domain.ArbeidsFordelingEnhet
 import no.nav.etterlatte.behandling.domain.SaksbehandlerEnhet
 import no.nav.etterlatte.behandling.klienter.NavAnsattKlient
+import no.nav.etterlatte.behandling.klienter.Norg2Klient
+import no.nav.etterlatte.common.Enheter
+import no.nav.etterlatte.common.IngenEnhetFunnetException
+import no.nav.etterlatte.common.IngenGeografiskOmraadeFunnetForEnhet
+import no.nav.etterlatte.common.klienter.PdlKlient
+import no.nav.etterlatte.libs.common.behandling.SakType
+import no.nav.etterlatte.libs.common.person.Folkeregisteridentifikator
+import org.slf4j.LoggerFactory
 
 interface EnhetService {
     suspend fun enheterForIdent(ident: String): List<SaksbehandlerEnhet>
@@ -10,13 +19,56 @@ interface EnhetService {
         ident: String,
         enhetId: String,
     ): Boolean
+
+    fun finnEnhetForPersonOgTema(
+        fnr: String,
+        tema: String,
+        saktype: SakType,
+    ): ArbeidsFordelingEnhet
 }
 
-class EnhetServiceImpl(val client: NavAnsattKlient) : EnhetService {
+class EnhetServiceImpl(
+    val client: NavAnsattKlient,
+    private val pdlKlient: PdlKlient,
+    private val norg2Klient: Norg2Klient,
+) : EnhetService {
+    private val logger = LoggerFactory.getLogger(this::class.java)
+
     override suspend fun enheterForIdent(ident: String) = client.hentSaksbehandlerEnhet(ident)
 
     override suspend fun harTilgangTilEnhet(
         ident: String,
         enhetId: String,
     ) = enheterForIdent(ident).any { enhet -> enhet.id == enhetId }
+
+    override fun finnEnhetForPersonOgTema(
+        fnr: String,
+        tema: String,
+        saktype: SakType,
+    ): ArbeidsFordelingEnhet {
+        val tilknytning = pdlKlient.hentGeografiskTilknytning(fnr, saktype)
+        val geografiskTilknytning = tilknytning.geografiskTilknytning()
+
+        return when {
+            tilknytning.ukjent ->
+                ArbeidsFordelingEnhet(
+                    Enheter.defaultEnhet.navn,
+                    Enheter.defaultEnhet.enhetNr,
+                )
+
+            geografiskTilknytning == null -> throw IngenGeografiskOmraadeFunnetForEnhet(
+                Folkeregisteridentifikator.of(fnr),
+                tema,
+            ).also {
+                logger.warn(it.message)
+            }
+
+            else -> finnEnhetForTemaOgOmraade(tema, geografiskTilknytning)
+        }
+    }
+
+    private fun finnEnhetForTemaOgOmraade(
+        tema: String,
+        omraade: String,
+    ) = norg2Klient.hentEnheterForOmraade(tema, omraade).firstOrNull() ?: throw IngenEnhetFunnetException(omraade, tema)
 }
